denta-server/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "denta-server"
version = "0.1.0"
description = "API backend for DentaApp — SaaS PWA para consultorios dentales"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "DentaApp Team"}
]

dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "supabase>=2.3",
    "python-jose[cryptography]>=3.3",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1",
    "httpx>=0.26",
    "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
